requirements.txt
# Core dependencies for SharePoint Graph API access
python-dotenv==1.0.0
requests==2.31.0
msal==1.24.1

# Microsoft Graph SDK (recommended approach)
msgraph-sdk==1.0.0
azure-identity==1.15.0

# Communication Assistant API dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
email-validator==2.1.0

# Database integration
supabase==1.0.4

# HTML processing and email formatting
jinja2==3.1.2
markdown==3.5.1
bleach==6.1.0

# Configuration
toml==0.10.2

# AI/ML dependencies
google-generativeai==0.8.5
mcp

# Monte Carlo simulation dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Optional: For additional functionality
# openpyxl==3.1.2  # If you need to read/write Excel files from SharePoint

 